requirements.txt
python-telegram-bot 13.15
urllib3             2.2.3
